Do not refer to the deprecated int-shift-by-long operations.
Shifts of non-`Long` lhs'es by `Long` rhs'es are deprecated in
Scala 2.13.0-M5. We avoid them while preserving behavior by
explicitly calling `.toInt` on the rhs.

+++ javalanglib/src/main/scala/java/lang/Character.scala
@@ -45,11 +45,11 @@ class Character(private val value: scala.Char)
   protected def +(x: String): String = value + x
 
   protected def <<(x: scala.Int): scala.Int = value << x
-  protected def <<(x: scala.Long): scala.Int = value << x
+  protected def <<(x: scala.Long): scala.Int = value << x.toInt
   protected def >>>(x: scala.Int): scala.Int = value >>> x
-  protected def >>>(x: scala.Long): scala.Int = value >>> x
+  protected def >>>(x: scala.Long): scala.Int = value >>> x.toInt
   protected def >>(x: scala.Int): scala.Int = value >> x
-  protected def >>(x: scala.Long): scala.Int = value >> x
+  protected def >>(x: scala.Long): scala.Int = value >> x.toInt
 
   protected def ==(x: scala.Byte): scala.Boolean = value == x
   protected def ==(x: scala.Short): scala.Boolean = value == x
